pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "yukishadow"
version = "0.1.0"
description = "Personal AI assistant and automation platform"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",

    # Configuration
    "pydantic-settings>=2.6.0",

    # Message bus
    "redis[hiredis]>=5.2.0",

    # Database
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",

    # Vector store
    "chromadb>=0.5.0",

    # Scheduling
    "apscheduler>=3.10.4",

    # Discord
    "discord.py>=2.4.0",

    # MCP
    "mcp>=1.0.0",

    # HTTP client (used by MCP server and Discord skill)
    "httpx>=0.28.0",

    # LLM providers (all optional at runtime – only the configured one is used)
    "ollama>=0.4.0",
    "openai>=1.56.0",
    "anthropic>=0.39.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
# Install on the Jetson Orin Nano
jetson = [
    "aiomqtt>=2.3.0",
]

# Install on the 5090 host when Jetson is connected
mqtt = [
    "aiomqtt>=2.3.0",
]

# Future: file watching
file_manager = [
    "watchdog>=5.0.0",
    "httpx>=0.28.0",
]

# Future: UI automation
ui_control = [
    "pyautogui>=0.9.54",
    "playwright>=1.48.0",
]

# Development
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[project.scripts]
yukishadow = "main:main"

[tool.hatch.build.targets.wheel]
packages = [
    "core",
    "orchestrator",
    "skills",
    "scheduler",
    "storage",
    "integrations",
    "mcp_server",
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
